refactor(client): subfolder and decompose domain radar components
The domains.$id route-private folder was a flat pile of three unrelated
components dominated by a 487-line -RadarChart.tsx. Sort them into themed
subfolders and break the radar chart into focused pieces so each file has one
job and the chart's data-shaping is separable from its rendering. This also
establishes the first nested *.-components/ subfolder pattern (routing-safe:
the - file prefix keeps files private; subfolders need no prefix).

- Group into radar/ (the RadarChart family) and topicLists/ (-TopicLinkList,
  -ExcludedTopicsList), each with an index.ts barrel.
- Decompose -RadarChart.tsx (487 -> 234 lines) into a slim orchestrator plus
  a -useRadarChartData hook (sorting/positioning useMemos), -RadarBackground
  (SVG grid), a unified -RadarStripSection (adopted + ignored strips via a
  mode prop), and -RadarLegendSidebar. All keep delegating to the shared
  @/components/radar primitives; behavior is unchanged.
- Add Storybook stories (no play functions) for every new component.
- Point the two domain route consumers at the new barrels.

Refs #473
