+++ packages/client/src/routes/domains.$id.-components/radar/index.ts
@@ -1,0 +1,4 @@
+// Public surface of the domain radar-chart family. The extracted sub-components
+// (-RadarBackground, -RadarStripSection, -RadarLegendSidebar, -useRadarChartData)
+// are internal to -RadarChart and intentionally not re-exported.
+export { RadarChart } from "./-RadarChart";
