+++ packages/client/src/routes/domains.$id.-components/topicLists/index.ts
@@ -1,0 +1,3 @@
+// Topic-list leaves shown on the domain overview page.
+export { ExcludedTopicsList } from "./-ExcludedTopicsList";
+export { TopicLinkList } from "./-TopicLinkList";
